style(desktop): 优化版本更新弹框样
- 根据左侧菜单栏状态动态调整更新弹框位置，避免遮挡登录用户信息。
- 缩小弹框高度并调整背景色，让版本更新提示更轻量、更贴合页面视觉。
